Implment an algorithm to prevent deadlocks
The Resque UI provides a button for deleting an entire queue. If that button
is pressed with jobs still in the queue, then the lock flags are orphaned in
the Reids datastore. These lock flags prevent those jobs from being re-added
to the queue.

This patch adds a timeout flag to the lock. After the timeout expires, the
lock can be cleared and jobs re-added to the queue.